Migrate CloneHearing
