+++ AdminWebsite/AdminWebsite.UnitTests/Controllers/HearingsController/PostHearingTests.cs
@@ -1,45 +1,26 @@
 using System.Linq;
 using AdminWebsite.Models;
-using AdminWebsite.Services;
 using AdminWebsite.UnitTests.Helper;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
 using System.Threading.Tasks;
 using AdminWebsite.Contracts.Requests;
 using BookingsApi.Client;
-using Autofac.Extras.Moq;
 using BookingsApi.Contract.V2.Enums;
 using BookingsApi.Contract.V2.Requests;
 using BookingsApi.Contract.V2.Responses;
-using VideoApi.Contract.Responses;
 using ParticipantRequest = AdminWebsite.Contracts.Requests.ParticipantRequest;
 
 namespace AdminWebsite.UnitTests.Controllers.HearingsController
 {
-    public class PostHearingTests
+    public class PostHearingTests : HearingsControllerTests
     {
-        private AutoMock _mocker;
-        private AdminWebsite.Controllers.HearingsController _controller;
-
         [SetUp]
-        public void Setup()
+        protected override void Setup()
         {
-            _mocker = AutoMock.GetLoose();
+            base.Setup();
 
-            _mocker.Mock<IConferenceDetailsService>().Setup(cs => cs.GetConferenceDetailsByHearingId(It.IsAny<Guid>(), false))
-                .ReturnsAsync(new ConferenceDetailsResponse
-                {
-                    MeetingRoom = new ()
-                    {
-                        AdminUri = "AdminUri",
-                        JudgeUri = "JudgeUri",
-                        ParticipantUri = "ParticipantUri",
-                        PexipNode = "PexipNode",
-                        PexipSelfTestNode = "PexipSelfTestNode",
-                        TelephoneConferenceId = "expected_conference_phone_id"
-                    }
-                });
-            _mocker.Mock<IBookingsApiClient>().Setup(bs => bs.GetHearingDetailsByIdV2Async(It.IsAny<Guid>()))
+            BookingsApiClient.Setup(bs => bs.GetHearingDetailsByIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(new HearingDetailsResponseV2()
                 {
                     Participants = new List<ParticipantResponseV2>
@@ -48,7 +29,6 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                     },
                     ServiceId = "Generic"
                 });
-            _controller = _mocker.Create<AdminWebsite.Controllers.HearingsController>();
         }
         
         [Test]
@@ -64,10 +44,10 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 BookingDetails = hearing
             };
             
-            _mocker.Mock<IBookingsApiClient>().Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
+            BookingsApiClient.Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
                 .Throws(ClientException.ForBookingsAPI(HttpStatusCode.InternalServerError));
 
-            var response = _controller.Post(bookingRequest);
+            var response = Controller.Post(bookingRequest);
 
             ((ObjectResult) response.Result.Result).StatusCode.Should().Be(500);
         }
@@ -85,10 +65,10 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 BookingDetails = hearing
             };
 
-            _mocker.Mock<IBookingsApiClient>().Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
+            BookingsApiClient.Setup(x => x.BookNewHearingWithCodeAsync(It.IsAny<BookNewHearingRequestV2>()))
                 .Throws(new Exception("Some internal error"));
             
-            var response = _controller.Post(bookingRequest);
+            var response = Controller.Post(bookingRequest);
 
             ((ObjectResult) response.Result.Result).StatusCode.Should().Be(500);
         }
@@ -108,19 +88,19 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 }
             };
 
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.GetHearingsByGroupIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(groupedHearings);
 
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.CloneHearingAsync(It.IsAny<Guid>(), It.IsAny<CloneHearingRequestV2>()))
                 .Verifiable();
 
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<NoContentResult>();
 
-            _mocker.Mock<IBookingsApiClient>().Verify(
+            BookingsApiClient.Verify(
                 x => x.CloneHearingAsync(It.IsAny<Guid>(), It.Is<CloneHearingRequestV2>(
                     y => y.ScheduledDuration == request.ScheduledDuration)),
                 Times.Exactly(1));
@@ -143,19 +123,19 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 });
             }
             
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.GetHearingsByGroupIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(groupedHearings);
 
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.CloneHearingAsync(It.IsAny<Guid>(), It.IsAny<CloneHearingRequestV2>()))
                 .Verifiable();
             
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<NoContentResult>();
             
-            _mocker.Mock<IBookingsApiClient>().Verify(
+            BookingsApiClient.Verify(
                 x => x.CloneHearingAsync(It.IsAny<Guid>(), It.IsAny<CloneHearingRequestV2>()),
                 Times.Exactly(1));
         }
@@ -168,20 +148,20 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
             var request = new MultiHearingRequest { StartDate = startDate, EndDate = endDate};
 
 
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
-            response.Should().BeOfType<BadRequestResult>();
+            response.Should().BeOfType<BadRequestObjectResult>();
         }
 
         [Test]
         public async Task Should_catch_BookingsApiException_by_clone_hearing()
         {
             var request = GetMultiHearingRequest();
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.CloneHearingAsync(It.IsAny<Guid>(), It.IsAny<CloneHearingRequestV2>()))
                 .Throws(new BookingsApiException("Error", (int)HttpStatusCode.BadRequest, "response", null, null));
 
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<BadRequestObjectResult>();
         }
@@ -190,11 +170,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
         public async Task Should_catch_InternalError_by_clone_hearing()
         {
             var request = GetMultiHearingRequest();
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.CloneHearingAsync(It.IsAny<Guid>(), It.IsAny<CloneHearingRequestV2>()))
                 .Throws(new BookingsApiException("Error", (int)HttpStatusCode.InternalServerError, "response", null, null));
 
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             ((ObjectResult) response).StatusCode.Should().Be(500);
         }
@@ -216,7 +196,7 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 }
             };
             
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.GetHearingsByGroupIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(groupedHearings);
             
@@ -226,11 +206,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 expectedDates.Add(date);
             }
             
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<NoContentResult>();
 
-            _mocker.Mock<IBookingsApiClient>().Verify(
+            BookingsApiClient.Verify(
                 x => x.CloneHearingAsync(It.IsAny<Guid>(), 
                     It.Is<CloneHearingRequestV2>(r => r.Dates.All(d => expectedDates.Contains(d)))),
                 Times.Exactly(1));
@@ -252,7 +232,7 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 }
             };
             
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.GetHearingsByGroupIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(groupedHearings);
 
@@ -265,11 +245,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 new(2022, 12, 20, 0, 0, 0, DateTimeKind.Utc)
             };
             
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<NoContentResult>();
 
-            _mocker.Mock<IBookingsApiClient>().Verify(
+            BookingsApiClient.Verify(
                 x => x.CloneHearingAsync(It.IsAny<Guid>(), 
                     It.Is<CloneHearingRequestV2>(r => r.Dates.All(d => expectedDates.Contains(d)))),
                 Times.Exactly(1));
@@ -295,7 +275,7 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 }
             };
             
-            _mocker.Mock<IBookingsApiClient>()
+            BookingsApiClient
                 .Setup(x => x.GetHearingsByGroupIdV2Async(It.IsAny<Guid>()))
                 .ReturnsAsync(groupedHearings);
 
@@ -306,11 +286,11 @@ namespace AdminWebsite.UnitTests.Controllers.HearingsController
                 new (2023, 1, 8, 0, 0, 0, DateTimeKind.Utc)
             };
             
-            var response = await _controller.CloneHearing(Guid.NewGuid(), request);
+            var response = await Controller.CloneHearing(Guid.NewGuid(), request);
 
             response.Should().BeOfType<NoContentResult>();
 
-            _mocker.Mock<IBookingsApiClient>().Verify(
+            BookingsApiClient.Verify(
                 x => x.CloneHearingAsync(It.IsAny<Guid>(), 
                     It.Is<CloneHearingRequestV2>(r => r.Dates.All(d => expectedDates.Contains(d)))),
                 Times.Exactly(1));
